Invalidate objects when owning context goes away.
When gcc_jit_context is destroyed, it frees memory allocated for types,
rvalues and other library objects. Meanwhile, Python objects
representing these types, rvalues and others are unaware that context is
gone, their _c_object pointer was freed and now points to an invalid
address.

This patch adds to every Object a weak reference to it's parent Context.
When Context is destroyed callback is called to clean up the dangling
pointer.